+++ src/plone/formwidget/recaptcha/upgrades.py
@@ -10,3 +10,13 @@ def install_browserlayer(context):
         run_dependencies=False,
         purge_old=False
     )
+
+
+def reapply_registry(context):
+    setup = getToolByName(context, 'portal_setup')
+    setup.runImportStepFromProfile(
+        'profile-plone.formwidget.recaptcha:default',
+        'plone.app.registry',
+        run_dependencies=False,
+        purge_old=False
+    )