+++ src/plone/formwidget/recaptcha/widget.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 from Acquisition import aq_inner
 from plone.formwidget.recaptcha.interfaces import IReCaptchaWidget
+from z3c.form import interfaces
 from z3c.form import widget
 from z3c.form.browser import text
 from zope.component import getMultiAdapter
+import zope.component
 import zope.interface
 import zope.schema.interfaces
 
@@ -29,6 +31,8 @@ class ReCaptchaWidget(text.TextWidget):
         return self.captcha.audio_url()
 
 
+@zope.component.adapter(zope.schema.interfaces.IField, interfaces.IFormLayer)
+@zope.interface.implementer(interfaces.IFieldWidget)
 def ReCaptchaFieldWidget(field, request):
     """IFieldWidget factory for CaptchaWidget."""
     return widget.FieldWidget(field, ReCaptchaWidget(request))